+++ src/Units/SpicyFryUnit.cpp
@@ -9,7 +9,7 @@ SpicyFryUnit::SpicyFryUnit(shared_ptr<TextLoader> textLoader, shared_ptr<EventMa
   this->radius= textLoader->getInteger(string("IDS_SF_AR"));;
   this->lunchMoney=textLoader->getInteger(string("IDS_SF_LM"));;
   this->typeID=textLoader->getTypeID(string("IDS_SFU"));
-  this->attackRate=textLoader->getInteger(string("IDS_SF_Attack_Rate"));
+  this->attackRate=textLoader->getDouble(string("IDS_SF_Attack_Rate"));
   this -> textures = textureLoader -> getTexture(typeID);
   this -> isAlly = true;
   //set the initial sprite texture
